refactor module, add histogram
