+++ pymc3/variational/__init__.py
@@ -23,3 +23,7 @@ from .inference import (
     FullRankADVI,
     fit,
 )
+from .approximations import Histogram
+
+from . import approximations
+from . import operators
